@@ -1,21 +1,25 @@
 {
-  "name": "gatsby-starter-forty-v2",
-  "description": "Gatsby Starter - Forty V2",
+  "name": "wbb-3.0",
+  "description": "We Build Black's newly redesigned website",
   "version": "1.0.0",
-  "author": "Hunter Chang",
+  "author": "Chloe Rice",
   "dependencies": {
+    "classnames": "^2.2.6",
     "gatsby": "^2.19.5",
     "gatsby-plugin-manifest": "^2.2.38",
     "gatsby-plugin-offline": "^3.0.32",
     "gatsby-plugin-react-helmet": "^3.1.21",
     "gatsby-plugin-sass": "^2.1.27",
     "node-sass": "^4.13.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
     "react-helmet": "^5.2.1"
   },
   "keywords": [
-    "gatsby"
+    "gatsby",
+    "forestry cms",
+    "react",
+    "we build black"
   ],
   "license": "MIT",
   "scripts": {
@@ -29,6 +33,6 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+    "url": "https://github.com/chloerice/wbb-3.0"
   }
 }
