@@ -1,18 +1,17 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { Link } from 'gatsby'
 
-const Header = (props) => (
+export default function Header({ onToggleMenu }) {
+  return (
     <header id="header" className="alt">
-        <Link to="/" className="logo"><strong>Forty</strong> <span>by HTML5 UP</span></Link>
-        <nav>
-            <a className="menu-link" onClick={props.onToggleMenu} href="javascript:;">Menu</a>
-        </nav>
+      <Link to="/" className="logo">
+        <strong>Forty</strong> <span>by HTML5 UP</span>
+      </Link>
+      <nav>
+        <a className="menu-link" onClick={onToggleMenu} href="javascript:;">
+          Menu
+        </a>
+      </nav>
     </header>
-)
-
-Header.propTypes = {
-    onToggleMenu: PropTypes.func
+  )
 }
-
-export default Header
