@@ -1,27 +1,48 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { Link } from 'gatsby'
 
-const Menu = (props) => (
+export default function Menu({ onToggleMenu }) {
+  return (
     <nav id="menu">
-        <div className="inner">
-            <ul className="links">
-                <li><Link onClick={props.onToggleMenu} to="/">Home</Link></li>
-                <li><Link onClick={props.onToggleMenu} to="/landing">Landing</Link></li>
-                <li><Link onClick={props.onToggleMenu} to="/generic">Generic</Link></li>
-                <li><Link onClick={props.onToggleMenu} to="/elements">Elements</Link></li>
-            </ul>
-            <ul className="actions vertical">
-                <li><a href="#" className="button special fit">Get Started</a></li>
-                <li><a href="#" className="button fit">Log In</a></li>
-            </ul>
-        </div>
-        <a className="close" onClick={props.onToggleMenu} href="javascript:;">Close</a>
+      <div className="inner">
+        <ul className="links">
+          <li>
+            <Link onClick={onToggleMenu} to="/">
+              Home
+            </Link>
+          </li>
+          <li>
+            <Link onClick={onToggleMenu} to="/landing">
+              Landing
+            </Link>
+          </li>
+          <li>
+            <Link onClick={onToggleMenu} to="/generic">
+              Generic
+            </Link>
+          </li>
+          <li>
+            <Link onClick={onToggleMenu} to="/elements">
+              Elements
+            </Link>
+          </li>
+        </ul>
+        <ul className="actions vertical">
+          <li>
+            <a href="#" className="button special fit">
+              Get Started
+            </a>
+          </li>
+          <li>
+            <a href="#" className="button fit">
+              Log In
+            </a>
+          </li>
+        </ul>
+      </div>
+      <a className="close" onClick={onToggleMenu} href="javascript:;">
+        Close
+      </a>
     </nav>
-)
-
-Menu.propTypes = {
-    onToggleMenu: PropTypes.func
+  )
 }
-
-export default Menu
