@@ -1,17 +1,17 @@
 import React from 'react'
 import Layout from '../components/layout'
 
-const NotFoundPage = () => (
-  <Layout>
-    <div id="main" className="alt">
+export default function NotFoundPage() {
+  return (
+    <Layout>
+      <div id="main" className="alt">
         <section id="one">
-            <div className="inner">
-                <h1>NOT FOUND</h1>
-                <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-            </div>
+          <div className="inner">
+            <h1>NOT FOUND</h1>
+            <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
+          </div>
         </section>
-    </div>
-  </Layout>
-)
-
-export default NotFoundPage
+      </div>
+    </Layout>
+  )
+}
